refactor: integrate Walker for file discovery and streamline configuration handling
