+++ crates/forge_app/src/lib.rs
@@ -14,7 +14,9 @@ mod tool_executor;
 mod tool_registry;
 mod truncation;
 mod utils;
+mod walker;
 
 pub use app::*;
 pub use fmt_output::*;
 pub use services::*;
+pub use walker::*;
