+++ crates/forge_app/src/services.rs
@@ -5,6 +5,8 @@ use forge_domain::{
     Environment, File, McpConfig, Model, ModelId, PatchOperation, ResultStream, Scope,
     ToolCallFull, ToolDefinition, ToolOutput, Workflow,
 };
+
+use crate::Walker;
 use merge::Merge;
 
 #[derive(Debug)]
@@ -180,7 +182,7 @@ pub trait WorkflowService {
 
 #[async_trait::async_trait]
 pub trait FileDiscoveryService: Send + Sync {
-    async fn collect(&self, max_depth: Option<usize>) -> anyhow::Result<Vec<File>>;
+    async fn collect_files(&self, config: Walker) -> anyhow::Result<Vec<File>>;
 }
 
 #[async_trait::async_trait]
@@ -425,8 +427,8 @@ impl<I: Services> WorkflowService for I {
 
 #[async_trait::async_trait]
 impl<I: Services> FileDiscoveryService for I {
-    async fn collect(&self, max_depth: Option<usize>) -> anyhow::Result<Vec<File>> {
-        self.file_discovery_service().collect(max_depth).await
+    async fn collect_files(&self, config: Walker) -> anyhow::Result<Vec<File>> {
+        self.file_discovery_service().collect_files(config).await
     }
 }
