+++ crates/forge_infra/src/forge_infra.rs
@@ -207,8 +207,8 @@ impl McpServerInfra for ForgeInfra {
 impl WalkerInfra for ForgeInfra {
     async fn walk(
         &self,
-        config: forge_services::WalkerConfig,
-    ) -> anyhow::Result<Vec<forge_services::WalkedFile>> {
+        config: forge_app::Walker,
+    ) -> anyhow::Result<Vec<forge_app::WalkedFile>> {
         self.walker_service.walk(config).await
     }
 }
